requirements.txt
torch >= 2.1.0
pandas >= 2.1.4
torchvision >= 0.16.2
pandas >= 2.1.4
transformers >= 4.36.2 #hugginface
numpy >= 1.26.3
scikit-learn >= 1.3.2
dvc >= 3.38.1
dvc-gdrive >= 3.0.1
hydra-core >= 1.3.2
accelerate >= 0.20.1
datasets >= 2.16.1
ruff >=0.1.11 #unsure if this is actually a requirement, however present in .toml